src/python/requirements.txt
# openai==0.28
python-dotenv>=0.19.0
requests>=2.26.0
torch
torchaudio
transformers
ffmpeg-python
numpy
librosa
pyannote-audio
pyannote.audio

langchain==0.3.15
langchain-openai==0.3.1
langchain-community==0.3.15
langchain-core==0.3.31
langchain-text-splitters==0.3.5
pandas==2.2.3
pinecone-client==5.0.1
pymongo==4.10.1
SQLAlchemy==2.0.37
